% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{area_relative_to_value}
\alias{area_relative_to_value}
\title{Determining Area Under the Distribution Curve of x, z, or bar x}
\usage{
area_relative_to_value(
  type = c("x", "z", "bar_x"),
  value_1,
  value_2 = NULL,
  mu = 0,
  sigma = 1,
  n = 1,
  direction = c("right", "left", "between")
)
}
\arguments{
\item{type}{A character string indicating the distribution under 
consideration, i.e., x, z, or bar x.}

\item{value_1}{The primary numeric index value for the distribution (required).}

\item{value_2}{The secondary numeric index value for the distribution (only 
required for the 'between' direction, default = NULL)}

\item{mu}{The population mean of the distribution (default = 0).}

\item{sigma}{The population standard deviation of the distribution (default = 1).}

\item{n}{The sample size (default = 1).}

\item{direction}{A character string indicating the directionality of the 
area relative to the index value(s).}
}
\value{
A printed output of area (as a percent) under the distribution 
curve for x, z, or bar x.
}
\description{
Calculate the area under the distribution curve to the left, to the right, 
or between index values of x, z, or bar x.
}
\examples{
output1 <- area_relative_to_value(type = "z", value_1 = -2.58, 
                                  direction = "left")
                                  
output2 <- area_relative_to_value(type = "x", value_1 = 3.56, 
                                  value_2 = 4.12, mu = 4.00,
                                  sigma = 0.25, direction = "between")
                                  
output3 <- area_relative_to_value(type = "barx", value_1 = 12,
                                  mu = 10, sigma = 2, n = 20,
                                  direction = "right")
}
