% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{value_relative_to_area}
\alias{value_relative_to_area}
\title{Determining Index Values of x or z Based on Area Under the Distribution 
Curve of x or z}
\usage{
value_relative_to_area(
  type = c("x", "z"),
  area,
  mu = 0,
  sigma = 1,
  direction = c("right", "left", "between")
)
}
\arguments{
\item{type}{A character string indicating the distribution under 
consideration, i.e., x or z.}

\item{area}{The numeric value for the area under the x or z distribution 
curve (as a proportion, i.e., 0.95 for 95\%).}

\item{mu}{The population mean of the distribution (default = 0).}

\item{sigma}{The population standard deviation of the distribution (default = 1).}

\item{direction}{A character string indicating the directionality of the 
area relative to the index value(s).}
}
\value{
A printed output of index values of x or z associated with the area 
under the distribution curve of x or z.
}
\description{
Calculate the index value(s) of x or z to the left, to the right, or 
bounding a given area under the distribution curve of x or z.
}
\examples{
output1 <- value_relative_to_area(type = "z", area = 0.95,
                                  direction = "left")

output2 <- value_relative_to_value(type = "x", area = 0.15, mu = 4.00,
                                   sigma = 0.25, direction = "between")
}
