% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{x2z}
\alias{x2z}
\title{Calculation of Standardized Z Scores From X Values}
\usage{
x2z(x, mu = 0, sigma = 1)
}
\arguments{
\item{x}{A vector of 1 or more numeric values of x.}

\item{mu}{The value of the population mean for the x distribution 
(0 = default value).}

\item{sigma}{The value of the population standard deviation for the x 
distribution (1 = default value).}
}
\value{
A printed output of one z score for each input value of x.
}
\description{
Convert numeric values of x from a non-standard normal distribution to z scores
from the standard normal distribution.
}
\note{
Use of mu = 0 and sigma = 1 (the default values) implies that the x 
distribution is equivalent to the standard normal distribution.  Thus,
the 'z' value returned under default conditions will be the input x value.
}
\examples{
z1 <- x2z(x = 3)

z2 <- x2z(x = 3, mu = 2.5, sigma = 0.4)

z3 <- x2z(x = c(10, 32, 112), mu = 55, sigma = 21)
}
