% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{z2x}
\alias{z2x}
\title{Calculation of Raw X Values From Standard Z Scores}
\usage{
z2x(z, mu = 0, sigma = 1)
}
\arguments{
\item{z}{A vector of 1 or more numeric values of z.}

\item{mu}{The value of the population mean for the x distribution 
(0 = default value).}

\item{sigma}{The value of the population standard deviation for the x 
distribution (1 = default value).}
}
\value{
A printed output of one x value for each input value of z.
}
\description{
Convert numeric values of z from the standard normal distribution to x values
from a non-standard normal distribution.
}
\note{
Use of mu = 0 and sigma = 1 (the default values) implies that the x 
distribution is equivalent to the standard normal distribution.  Thus,
the 'x' value returned under default conditions will be the input z value.
}
\examples{
x1 <- z2x(z = -3.00)

x2 <- z2x(z = -3.00, mu = 5, sigma = 2)

x3 <- z2x(c(-1.96, 0.00, 2.56), mu = 3.34, sigma = 1.67)
}
